Add 2 Notification buttons logic and notification fetch via API
The fronend now fetches the backend periodically to know if there are
new notifications and, in that case, shows the notification button below
the last bubble.

Tha button, when pressed, will call the ReadNotification intent (so,
it's the same if the user enters read notifications text) and print the
notification with the highest priority.

The NotificaiontNumber API is used to display and load the number of new
notifications. Currently, it uses a polling approach to fetch
periodically the new notifications but this is just for testing purposes
and will be changed in the future.

Finally, now via the frontend it's possible to CREATE (via API) new
notifications, that will be sent to the nodejs backend and shown in the
frontend. While configuring the testing buttons it's possible to specify
the appliance, the type and the priority of the notification that should
be generated.